backend/requirements.txt
fastapi
uvicorn
motor
tavily-python
google-generativeai
python-dotenv
